Fix config serialization

+++ src/config.h
@@ -17,7 +17,7 @@ struct Config {
     std::vector<int> fovPresets = { 30, 45, 60, 75, 90, 110 };
 
     bool interpolate = true;
-    float smoothing = 0.2;
+    double smoothing = 0.2;
 
     int hookKey = VK_UP;
     int enableKey = VK_DOWN;
